ScrollboxManager: Refactor 3
